Split native Zenmap profile storage helpers
